kernel_compile: Add R-Car linux-next support
This does not use the officially recommended compiler as the support is
indended for testing and demonstration purposes only. We can use the
existing generic linaro AARCH64 cross compiler used for the TX2 as a
result. To do this we will also have to account for the vendor specific
sub directories in the arm64 tree.

Signed-off-by: Gareth Williams <garethwilliams857@gmail.com>