kernel_compile: Implement platform generic solution
Parsing the platforms directory for the names of supported platforms and
using a script to hold the meta information on each platform allows the
kernel compilation script itself to be implemented generically. This is
at the cost of the cross compiler downloading feature but allows for a
more scalable solution.

Signed-off-by: Gareth Williams <garethwilliams857@gmail.com>